Replace switch with lookup for unicode props
Previously, we code-generated an enormous switch statement
with a branch for each unicode property range. This worked
but was slow to compile, especially for line break properties.
Additionally, the large switch statement triggered an error
in the dlv debugger.

Instead, generate lookup tables. For ASCII, we can do a direct
lookup in an array, and for non-ASCII we can binary search
the ranges. This is just as fast for the ASCII case, and
is much faster to compile.