Refactor word commands
The main change is to use a single
locator function for word objects that returns both the start
and end position. This simplifies the logic and fixes a few edge
cases with leading/trailing whitespace. It should also make it
easier to extend the locators to support counting multiple words.