Move fuzzy search lexicographic order sort
Previously, the sort for emptyQueryShowAll occurred in fuzzy.Index.
Moving it to state.ShowMenu allows us to sort in a different order
for the "parent directories" command, where we want to list paths
in descending order by length (reverse lexicographic). This also avoids
repeated sorts if the user types a query then deletes it.