Insert after cursor from shell cmd with no selection
When inserting text using a shellcmd without any text
selected, insert the text *after* the cursor position
instead of before.

This is more intuitive when inserting text at the end
of a line. If the user types a space, then escapes to normal
mode, the cursor moves "back" to the space. If the user then
executes a menu command to insert some text, the text should
be inserted *after* the space, not before.