Allow undo to recover changes from the redo log
If the user uses "undo" to revert changes, then edits
the document, the reverted changes are discarded.
This matches vim's behavior, but can confuse users.

A better approach is presented in
"Resolving the Great Undo-Redo Quandary"
https://github.com/zaboople/klonk/blob/master/TheGURQ.md

Instead of discarding changes in the redo log, add their
inverses to the undo log. Then the user can "undo" back
to the state with the redo changes applied.