Support bracketed paste
Bracketed paste is a terminal feature that allows TUI applications
to detect the start and end of pasted text. Previously, we were
interpreting pasted text as user key presses, which caused the following
problems:

* Pasting was very slow, especially with syntax highlighting, since each character
  was processed separately.

* Auto-indent would add indentation to pasted text.

* Normal mode would interpret pasted text as commands, which could mess up the document.

Bracketed paste solves these problems. Since aretext knows the beginning/end of the
pasted text, it can buffer it then process the full text in a single action,
which is significantly faster. Additionally, aretext can skip auto-indent
and normal mode commands for pasted text, avoiding the edge cases of the previous
approach.