Remove github.com/google/shlex
The package has been archived in Github for several years.
We were using it to split the words in the $SHELL environment variable.

I believe this was incorrect behavior, since splitting a string
into command arguments depends on the syntax of the shell.

Usually, $SHELL is set to the value of the user's login shell
(as stored in /etc/passwd), which is interpreted as an executable.
The user could change this after the shell starts, but it seems
highly unlikely anyone would do this.