Implement delete search object commands
Implement "d/" and "d?" commands to delete to a search match.

This is the first time a search mode command can modify the document,
so it required some changes to how we checkpoint the undo log in search mode.

This also required a bit of a hack to replay the delete action using
the exact query used in the original command.