Show original cursor position while in search mode
Previously, search mode would move the cursor to the search query
at the bottom of the screen, with no indication of the original
cursor position in the document. This is annoying for commands
like "d/" because the user can't see the range of text that
will be deleted.

Fix it by displaying the original cursor position in the document
using a reverse+dim style. This style should be supported by nearly
every terminal emulator and won't conflict with syntax highlighting
colors.