Cross-platform file path check in parent directory menu
The parent directory menu contains an entry for each
of the ancestors of the current working directory.
Previously, this was generated in a Linux-specific
way by checking for "/" or "." paths to terminate
the loop. On Windows, this would trigger an infinite
loop and eventually crash the program with an out-of-memory
error.

Fortunately, there is a better way to detect when
we have reached the root directory. The comments
for filepath.Clean and filepath.Dir say that
the returned path does not end in a separator unless it is
the root directory. So we can check for a trailing separator
in the loop condition.

Signed-off-by: Will Daly <will.e.daly@gmail.com>