Fix change-inner-block behavior for single-line blocks
Previously, aretext would insert newlines after the start
delimiter and before the end delimiter even for single-line
blocks.

For example, this:

    { xyz }

... would become:

    {
    }

This differs from vim's behavior, which does not insert
the newlines.

Fix it by preserving the newlines immediately after the start
delimiter and immediately before the end delimiter, and stop
inserting newlines.