Skip markdown setext heading for single hyphen
A single hyphen after a paragraph is technically a setext heading
according to the commonmark spec. However, it's really annoying
that text gets briefly highlighted as setext when the user has
started typing the first item in a list.

Fix it by treating '-' after a paragraph as part of the paragraph
rather than a setext heading underline.