Abort save if file was moved or deleted
Previously, if a file was moved or deleted outside the editor
and a user saved the file, aretext would recreate the file at
the original path without any warning.

With this change, aretext now detects when a file was moved/deleted
and shows an error. The user can still use "force save" to save the
file at the original path.