Add HidePatterns configuration, deprecate HideDirectories
The current configuration allows hiding directories, but not hiding
files with a particular extension (".obj", ".png", etc.).

Add a new config HidePatterns that uses glob patterns to hide
either a file or directory.

This introduces a slight change in behavior where HideDirectories
can now exclude both a file and a directory (it becomes an alias
for HidePatterns).

HideDirectories is marked as deprecated in the docs, but will
be kept for backwards compatibility.