Remove clipboard interaction from insert shell cmd
Previously, inserting the output of a shell cmd involved
writing the output to a special clipboard page, then pasting
from the clipboard into the document.

This is just silly. Worse, it stores an unnecessary separate
copy of the the shell cmd output.

Fix it by inserting the shell cmd output directly into the document.