Upgrade tcell from 2.9.0 to 2.12.2
The biggest change is that tcell has replaced go-runewidth with
rivo/uniseg for determining cellwidth. Switching to uniseg simplified
a lot of the cellwidth workarounds, at the cost of an extra conversion
from []rune to string for each grapheme cluster.

There were also some unexpected changes to the fill character used
by SimulationScreen ('X' instead of ' '). I'm not sure from the tcell
commit history what caused it, but adding a Clear() to the unit tests
restored the previous behavior.