Fix incorrect grapheme cluster breaking in buffer display
The loop in `drawLineAndSetCursor` wasn't correctly resetting
the grapheme cluster breaker. This could cause the same rune
to be processed twice and incorrectly introduce a break within
a grapheme cluster.

For example, the string " \U0001F1FA\U0001F1F8" is a space
followed by the flag for the USA (two runes). In the first pass
through the loop, the " " and "\U0001F1FA" runes are processed,
and a grapheme cluster break is detected after the space.
In the second iteration of the loop, "\U0001F1FA" is processed
again, then "\U0001F1F8". The sequence is misinterpreted as
"RI + RI + RI" instead of "RI + RI", so a GC break is incorrectly
inserted between the two runes.

In the terminal, this would appear as inconsistent spacing and
"dirty" leftover characters.