Display space with combining mark modifier
The unicode grapheme cluster segmentation algorithm rule GB9
prevents a break between any codepoint followed by a codepoint
with the "Extend" property, even if the first codepoint is a
space.

Previously the display code would ignore any runes following
a space character; now it displays them.

Tabs continue to hide modifiers for now. This simplifies the
implementation, and I'm not aware of any use cases for supporting it.